requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
langchain==0.1.9
transformers==4.37.2
python-dotenv==1.0.1
pydantic==2.6.1
spacy==3.7.2
nltk==3.8.1
torch==2.2.0
tensorflow==2.15.0
requests==2.31.0
python-multipart==0.0.9
certifi==2025.1.31
